TS-38561 Refactor the Artifactory inner class "CommitInfo" to its own class
