+++ agent/src/main/java/com/teamscale/jacoco/agent/upload/artifactory/ArtifactoryConfig.java
@@ -1,66 +1,62 @@
 package com.teamscale.jacoco.agent.upload.artifactory;
 
-import java.io.File;
-import java.io.IOException;
-import java.time.ZonedDateTime;
-import java.time.format.DateTimeFormatter;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Objects;
-import java.util.Properties;
-
-import org.conqat.lib.commons.collections.Pair;
-
 import com.teamscale.client.CommitDescriptor;
 import com.teamscale.client.StringUtils;
+import com.teamscale.jacoco.agent.commit_resolution.git_properties.CommitInfo;
 import com.teamscale.jacoco.agent.commit_resolution.git_properties.GitPropertiesLocatorUtils;
 import com.teamscale.jacoco.agent.commit_resolution.git_properties.GitSingleProjectPropertiesLocator;
 import com.teamscale.jacoco.agent.commit_resolution.git_properties.InvalidGitPropertiesException;
 import com.teamscale.jacoco.agent.options.AgentOptionParseException;
 import com.teamscale.jacoco.agent.options.AgentOptionsParser;
 import com.teamscale.jacoco.agent.options.FilePatternResolver;
-
 import okhttp3.HttpUrl;
+import org.conqat.lib.commons.collections.Pair;
+
+import java.io.File;
+import java.io.IOException;
+import java.time.ZonedDateTime;
+import java.time.format.DateTimeFormatter;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Properties;
 
 /** Config necessary to upload files to an azure file storage. */
 public class ArtifactoryConfig {
 	/**
-	 * Option to specify the artifactory URL. This shall be the entire path down to
-	 * the directory to which the coverage should be uploaded to, not only the base
-	 * url of artifactory.
+	 * Option to specify the artifactory URL. This shall be the entire path down to the directory to which the coverage
+	 * should be uploaded to, not only the base url of artifactory.
 	 */
 	public static final String ARTIFACTORY_URL_OPTION = "artifactory-url";
 
 	/**
-	 * Username that shall be used for basic auth. Alternative to basic auth is to
-	 * use an API key with the {@link ArtifactoryConfig#ARTIFACTORY_API_KEY_OPTION}
+	 * Username that shall be used for basic auth. Alternative to basic auth is to use an API key with the
+	 * {@link ArtifactoryConfig#ARTIFACTORY_API_KEY_OPTION}
 	 */
 	public static final String ARTIFACTORY_USER_OPTION = "artifactory-user";
 
 	/**
-	 * Password that shall be used for basic auth. Alternative to basic auth is to
-	 * use an API key with the {@link ArtifactoryConfig#ARTIFACTORY_API_KEY_OPTION}
+	 * Password that shall be used for basic auth. Alternative to basic auth is to use an API key with the
+	 * {@link ArtifactoryConfig#ARTIFACTORY_API_KEY_OPTION}
 	 */
 	public static final String ARTIFACTORY_PASSWORD_OPTION = "artifactory-password";
 
 	/**
 	 * API key that shall be used to authenticate requests to artifactory with the
-	 * {@link com.teamscale.jacoco.agent.upload.artifactory.ArtifactoryUploader#ARTIFACTORY_API_HEADER}.
-	 * Alternatively basic auth with username
-	 * ({@link ArtifactoryConfig#ARTIFACTORY_USER_OPTION}) and password
+	 * {@link com.teamscale.jacoco.agent.upload.artifactory.ArtifactoryUploader#ARTIFACTORY_API_HEADER}. Alternatively
+	 * basic auth with username ({@link ArtifactoryConfig#ARTIFACTORY_USER_OPTION}) and password
 	 * ({@link ArtifactoryConfig#ARTIFACTORY_PASSWORD_OPTION}) can be used.
 	 */
 	public static final String ARTIFACTORY_API_KEY_OPTION = "artifactory-api-key";
 
 	/**
-	 * Option that specifies if the legacy path for uploading files to artifactory
-	 * should be used instead of the new standard path.
+	 * Option that specifies if the legacy path for uploading files to artifactory should be used instead of the new
+	 * standard path.
 	 */
 	public static final String ARTIFACTORY_LEGACY_PATH_OPTION = "artifactory-legacy-path";
 
 	/**
-	 * Option that specifies under which path the coverage file shall lie within the
-	 * zip file that is created for the upload.
+	 * Option that specifies under which path the coverage file shall lie within the zip file that is created for the
+	 * upload.
 	 */
 	public static final String ARTIFACTORY_ZIP_PATH_OPTION = "artifactory-zip-path";
 
@@ -70,14 +66,12 @@ public class ArtifactoryConfig {
 	public static final String ARTIFACTORY_PATH_SUFFIX = "artifactory-path-suffix";
 
 	/**
-	 * Specifies the location of the JAR file which includes the git.properties
-	 * file.
+	 * Specifies the location of the JAR file which includes the git.properties file.
 	 */
 	public static final String ARTIFACTORY_GIT_PROPERTIES_JAR_OPTION = "artifactory-git-properties-jar";
 
 	/**
-	 * Specifies the date format in which the commit timestamp in the git.properties
-	 * file is formatted.
+	 * Specifies the date format in which the commit timestamp in the git.properties file is formatted.
 	 */
 	public static final String ARTIFACTORY_GIT_PROPERTIES_COMMIT_DATE_FORMAT_OPTION = "artifactory-git-properties-commit-date-format";
 
@@ -108,15 +102,13 @@ public class ArtifactoryConfig {
 	public CommitInfo commitInfo;
 
 	/**
-	 * We save the original dateTimeFormatter pattern, as it would be lost after
-	 * creating the formatter otherwise. Can be reused when retrying unsuccessful
-	 * uploads.
+	 * We save the original dateTimeFormatter pattern, as it would be lost after creating the formatter otherwise. Can
+	 * be reused when retrying unsuccessful uploads.
 	 */
 	public String dateTimeFormatterPattern = "yyyy-MM-dd'T'HH:mm:ssZ";
 
 	/**
-	 * Related to
-	 * {@link ArtifactoryConfig#ARTIFACTORY_GIT_PROPERTIES_COMMIT_DATE_FORMAT_OPTION}
+	 * Related to {@link ArtifactoryConfig#ARTIFACTORY_GIT_PROPERTIES_COMMIT_DATE_FORMAT_OPTION}
 	 */
 	public DateTimeFormatter gitPropertiesCommitTimeFormat = DateTimeFormatter.ofPattern(dateTimeFormatterPattern);
 
@@ -132,42 +124,42 @@ public class ArtifactoryConfig {
 	 * @return true if it has successfully process the given option.
 	 */
 	public static boolean handleArtifactoryOptions(ArtifactoryConfig options, FilePatternResolver filePatternResolver,
-			String key, String value) throws AgentOptionParseException {
+												   String key, String value) throws AgentOptionParseException {
 		switch (key) {
-		case ARTIFACTORY_URL_OPTION:
-			options.url = AgentOptionsParser.parseUrl(key, value);
-			return true;
-		case ARTIFACTORY_USER_OPTION:
-			options.user = value;
-			return true;
-		case ARTIFACTORY_PASSWORD_OPTION:
-			options.password = value;
-			return true;
-		case ARTIFACTORY_LEGACY_PATH_OPTION:
-			options.legacyPath = Boolean.parseBoolean(value);
-			return true;
-		case ARTIFACTORY_ZIP_PATH_OPTION:
-			options.zipPath = StringUtils.stripSuffix(value, "/");
-			return true;
-		case ARTIFACTORY_PATH_SUFFIX:
-			options.pathSuffix = StringUtils.stripSuffix(value, "/");
-			return true;
-		case ARTIFACTORY_GIT_PROPERTIES_JAR_OPTION:
-			options.commitInfo = ArtifactoryConfig.parseGitProperties(filePatternResolver,
-					options.gitPropertiesCommitTimeFormat, key, value);
-			return true;
-		case ARTIFACTORY_GIT_PROPERTIES_COMMIT_DATE_FORMAT_OPTION:
-			options.gitPropertiesCommitTimeFormat = DateTimeFormatter.ofPattern(value);
-			options.dateTimeFormatterPattern = value;
-			return true;
-		case ARTIFACTORY_API_KEY_OPTION:
-			options.apiKey = value;
-			return true;
-		case ARTIFACTORY_PARTITION:
-			options.partition = value;
-			return true;
-		default:
-			return false;
+			case ARTIFACTORY_URL_OPTION:
+				options.url = AgentOptionsParser.parseUrl(key, value);
+				return true;
+			case ARTIFACTORY_USER_OPTION:
+				options.user = value;
+				return true;
+			case ARTIFACTORY_PASSWORD_OPTION:
+				options.password = value;
+				return true;
+			case ARTIFACTORY_LEGACY_PATH_OPTION:
+				options.legacyPath = Boolean.parseBoolean(value);
+				return true;
+			case ARTIFACTORY_ZIP_PATH_OPTION:
+				options.zipPath = StringUtils.stripSuffix(value, "/");
+				return true;
+			case ARTIFACTORY_PATH_SUFFIX:
+				options.pathSuffix = StringUtils.stripSuffix(value, "/");
+				return true;
+			case ARTIFACTORY_GIT_PROPERTIES_JAR_OPTION:
+				options.commitInfo = ArtifactoryConfig.parseGitProperties(filePatternResolver,
+						options.gitPropertiesCommitTimeFormat, key, value);
+				return true;
+			case ARTIFACTORY_GIT_PROPERTIES_COMMIT_DATE_FORMAT_OPTION:
+				options.gitPropertiesCommitTimeFormat = DateTimeFormatter.ofPattern(value);
+				options.dateTimeFormatterPattern = value;
+				return true;
+			case ARTIFACTORY_API_KEY_OPTION:
+				options.apiKey = value;
+				return true;
+			case ARTIFACTORY_PARTITION:
+				options.partition = value;
+				return true;
+			default:
+				return false;
 		}
 	}
 
@@ -190,7 +182,8 @@ public class ArtifactoryConfig {
 
 	/** Parses the commit information form a git.properties file. */
 	public static CommitInfo parseGitProperties(FilePatternResolver filePatternResolver,
-			DateTimeFormatter gitPropertiesCommitTimeFormat, String optionName, String value)
+												DateTimeFormatter gitPropertiesCommitTimeFormat, String optionName,
+												String value)
 			throws AgentOptionParseException {
 		File jarFile = filePatternResolver.parsePath(optionName, value).toFile();
 		try {
@@ -215,7 +208,8 @@ public class ArtifactoryConfig {
 
 	/** Parses the commit information from a git.properties file. */
 	public static List<CommitInfo> parseGitProperties(File file, boolean isJarFile,
-			DateTimeFormatter gitPropertiesCommitTimeFormat, boolean recursiveSearch)
+													  DateTimeFormatter gitPropertiesCommitTimeFormat,
+													  boolean recursiveSearch)
 			throws IOException, InvalidGitPropertiesException {
 		List<Pair<String, Properties>> entriesWithProperties = GitPropertiesLocatorUtils.findGitPropertiesInFile(file,
 				isJarFile, recursiveSearch);
@@ -229,7 +223,7 @@ public class ArtifactoryConfig {
 			String branchName = getGitPropertiesValue(properties,
 					GitSingleProjectPropertiesLocator.GIT_PROPERTIES_GIT_BRANCH, entry, file);
 			long timestamp = ZonedDateTime.parse(getGitPropertiesValue(properties,
-					GitSingleProjectPropertiesLocator.GIT_PROPERTIES_GIT_COMMIT_TIME, entry, file),
+							GitSingleProjectPropertiesLocator.GIT_PROPERTIES_GIT_COMMIT_TIME, entry, file),
 					gitPropertiesCommitTimeFormat).toInstant().toEpochMilli();
 			result.add(new CommitInfo(revision, new CommitDescriptor(branchName, timestamp)));
 
@@ -252,40 +246,4 @@ public class ArtifactoryConfig {
 		return value;
 	}
 
-	/** Hold information regarding a commit. */
-	public static class CommitInfo {
-		/** The revision information (git hash). */
-		public String revision;
-
-		/** The commit descriptor. */
-		public CommitDescriptor commit;
-
-		/** Constructor. */
-		public CommitInfo(String revision, CommitDescriptor commit) {
-			this.revision = revision;
-			this.commit = commit;
-		}
-
-		@Override
-		public String toString() {
-			return commit + "/" + revision;
-		}
-
-		@Override
-		public boolean equals(Object o) {
-			if (this == o) {
-				return true;
-			}
-			if (o == null || getClass() != o.getClass()) {
-				return false;
-			}
-			CommitInfo that = (CommitInfo) o;
-			return Objects.equals(revision, that.revision) && Objects.equals(commit, that.commit);
-		}
-
-		@Override
-		public int hashCode() {
-			return Objects.hash(revision, commit);
-		}
-	}
 }
